+++ src/core/server/effects/vehicleRepair.ts
@@ -1,11 +1,18 @@
 import * as alt from 'alt-server';
+import { Vehicle_Door_List } from '../../shared/enums/vehicle';
 import { AnimationFlags } from '../../shared/flags/animation';
 import { Item } from '../../shared/interfaces/Item';
+import { Task, TaskCallback } from '../../shared/interfaces/TaskTimeline';
 import { playerFuncs } from '../extensions/Player';
 import { vehicleFuncs } from '../extensions/Vehicle';
 import { getForwardVector } from '../utility/vector';
 
-alt.on('effect:Vehicle:Repair', (player: alt.Player, item: Item, slot: number, tab: number) => {
+const isUsingTimeline: Array<{ player: alt.Player; vehicle: alt.Vehicle }> = [];
+
+alt.onClient('task:Vehicle:Repair:Timeline', handleRepairTimeline);
+alt.on('effect:Vehicle:Repair', handleRepair);
+
+function handleRepair(player: alt.Player) {
     const closestVehicle = playerFuncs.utility.getVehicleInFrontOf(player, 2);
 
     if (!closestVehicle) {
@@ -13,27 +20,9 @@ alt.on('effect:Vehicle:Repair', (player: alt.Player, item: Item, slot: number, t
         return;
     }
 
-    let passedAllTests = true;
-
-    // In Inventory
-    if (tab !== null && tab !== undefined) {
-        passedAllTests = playerFuncs.inventory.inventoryRemove(player, slot, tab);
-        playerFuncs.sync.inventory(player);
-        playerFuncs.save.field(player, 'inventory', player.data.inventory);
-    } else {
-        // In Toolbar
-        const currentItem = playerFuncs.inventory.getToolbarItem(player, item.slot);
-        if (!currentItem) {
-            playerFuncs.emit.message(player, `Could not find item.`);
-            return;
-        }
-
-        passedAllTests = playerFuncs.inventory.toolbarRemove(player, currentItem.slot);
-        playerFuncs.sync.inventory(player);
-        playerFuncs.save.field(player, 'toolbar', player.data.toolbar);
-    }
-
-    if (!passedAllTests) {
+    const removedKit = playerFuncs.inventory.findAndRemove(player, 'Repair Kit');
+    if (!removedKit) {
+        playerFuncs.emit.message(player, `You do not have a repair kit.`);
         return;
     }
 
@@ -44,19 +33,52 @@ alt.on('effect:Vehicle:Repair', (player: alt.Player, item: Item, slot: number, t
         z: closestVehicle.pos.z
     };
 
-    playerFuncs.emit.moveTo(player, fwdPosition as alt.Vector3);
+    const timeline: Array<Task | TaskCallback> = [
+        {
+            // taskGoToCoordAnyMeans(ped: number, x: number, y: number, z: number, speed: number, p5: any, p6: boolean, walkingStyle: number, p8: number
+            nativeName: 'taskGoToCoordAnyMeans',
+            params: [fwdPosition.x, fwdPosition.y, fwdPosition.z, 1, 0, false, 786603, 0],
+            timeToWaitInMs: 5000
+        },
+        {
+            // taskTurnPedToFaceCoord(ped: number, x: number, y: number, z: number, duration: number)
+            nativeName: 'taskTurnPedToFaceCoord',
+            params: [closestVehicle.pos.x, closestVehicle.pos.y, closestVehicle.pos.z, 2000],
+            timeToWaitInMs: 2000
+        },
+        {
+            callbackName: 'task:Vehicle:Repair:Timeline'
+        }
+    ];
 
-    alt.setTimeout(() => {
-        playerFuncs.emit.animation(
-            player,
-            'mp_car_bomb',
-            'car_bomb_mechanic',
-            AnimationFlags.NORMAL | AnimationFlags.REPEAT,
-            12000
-        );
-    }, 2000);
+    isUsingTimeline.push({ player, vehicle: closestVehicle });
+    playerFuncs.emit.taskTimeline(player, timeline);
+}
+
+function handleRepairTimeline(player: alt.Player) {
+    const index = isUsingTimeline.findIndex((data) => data.player === player);
+    if (index <= -1) {
+        return;
+    }
+
+    const closestVehicle = isUsingTimeline[index].vehicle;
+    isUsingTimeline.splice(index, 1);
+
+    if (!closestVehicle || !closestVehicle.valid) {
+        return;
+    }
+
+    playerFuncs.emit.animation(
+        player,
+        'mp_car_bomb',
+        'car_bomb_mechanic',
+        AnimationFlags.NORMAL | AnimationFlags.REPEAT,
+        12000
+    );
+
+    vehicleFuncs.setter.doorOpen(closestVehicle, player, Vehicle_Door_List.HOOD, true, true);
 
     alt.setTimeout(() => {
         vehicleFuncs.utility.repair(closestVehicle);
     }, 12000);
-});
+}