+++ src/core/server/extensions/vehicleFunctions/setter.ts
@@ -23,9 +23,17 @@ function lock(v: alt.Vehicle, player: alt.Player, lockState: Vehicle_Lock_State)
     return true;
 }
 
-function doorOpen(v: alt.Vehicle, player: alt.Player, index: Vehicle_Door_List, state: boolean): void {
-    if (!getter.isOwner(v, player) && !keys.has(v, player) && getter.lockState(v) !== Vehicle_Lock_State.UNLOCKED) {
-        return;
+function doorOpen(
+    v: alt.Vehicle,
+    player: alt.Player,
+    index: Vehicle_Door_List,
+    state: boolean,
+    bypass: boolean = false
+): void {
+    if (!bypass) {
+        if (!getter.isOwner(v, player) && !keys.has(v, player) && getter.lockState(v) !== Vehicle_Lock_State.UNLOCKED) {
+            return;
+        }
     }
 
     // alt.log(`${doorName}: ${oppositeValue}`);