+++ src/core/shared/enums/system.ts
@@ -18,6 +18,7 @@ export enum SYSTEM_EVENTS {
     PLAYER_EMIT_FRONTEND_SOUND = 'sound:Frontend',
     PLAYER_EMIT_NOTIFICATION = 'notification:Show',
     PLAYER_EMIT_TASK_MOVE = 'task:Move',
+    PLAYER_EMIT_TASK_TIMELINE = 'task:Timeline',
     //
     PLAYER_RELOAD = 'player:ForceReload',
     //