+++ documentation/changelog.md
@@ -46,6 +46,9 @@ description: List of all the changes and current versions.
 + Added YouTube Audio Streaming Service with Distance
 + Added Burgular Alarm to Heist Job
 + Repair Kit Basic for Vehicle
++ Added Task Timelines to Server Side
++ Timelines allow you to call a set of native player tasks to perform actions in order.
++ Upgrade Vehicle Repair
 ```
 
 ## 1.5.0 - January 28-31, 2021